rename WrappingError to WrappedError
